Notification Implementation
Notification can be sent to the device now.
The backend could retrieve weather data now.
The device will keep the user's login state until the sign-out button is clicked.
When the user sign-out the device is removed from the backend database.
The server now has a scheduled notification.
Still need to check plant minimum data to compare with the current weather data before sending the notification.